Rename sort_iati.py to sort_xml.py
